Logging all steps
Extent Report logs all steps, including passed checks and failures. Need to add screenshots and categories.
