+++ src/test/java/autoFramework/AutoLogger.java
@@ -12,7 +12,7 @@ public class AutoLogger {
 
     private final TestContextLogger testContextLogger = new TestContextLogger();
     private final TestExecutionContext testExecutionContext = new TestExecutionContext();
-
+    private final ExtentLogger extentLogger = new ExtentLogger();
     // Needs to be static for Listeners
     private static int stepNumber = 1;
 
@@ -50,7 +50,7 @@ public class AutoLogger {
     public void Info(String message)
     {
         testContextLogger.Info(message);
-        //ExtentFactory.getInstance().Info(message);
+        extentLogger.Info(message);
     }
 
     /** Log setup, teardown, or API info. NOTE: Logging this info for tests
@@ -78,7 +78,7 @@ public class AutoLogger {
     {
         message = "   (PASS)   Got : " + message;
         testContextLogger.Pass(message);
-        //ExtentFactory.getInstance().Pass(message);
+        extentLogger.PassStep(message);
     }
 
     public void FailCompare(Throwable ex)
@@ -95,14 +95,14 @@ public class AutoLogger {
 
         testContextLogger.Fail("   (FAIL) " + actual);
         testContextLogger.Fail(expected);
-        //ExtentFactory.getInstance().Fail(actual, expected);
+        extentLogger.Fail(actual, expected);
     }
 
     /** Log what is currently being done by the code. Increments int stepNumber.*/
     public void Step(String message)
     {
         testContextLogger.Step(message,stepNumber);
-        //ExtentFactory.getInstance().LogStep(message, stepNumber);
+        extentLogger.LogStep(message, stepNumber);
 
         stepNumber++;
     }
@@ -114,7 +114,7 @@ public class AutoLogger {
     {
         message = MessageFormat.format(message, actual);
         testContextLogger.Step(message,stepNumber);
-        //ExtentFactory.getInstance().LogStep(message, stepNumber);
+        extentLogger.LogStep(message, stepNumber);
 
         stepNumber++;
     }
@@ -127,7 +127,7 @@ public class AutoLogger {
         SetUpInfo("REQUEST BODY: " +  requestBody);
         SetUpInfo("STATUS CODE: " + response.statusCode());
         SetUpInfo("RESPONSE BODY: " + response.asString());
-        //ExtentFactory.getInstance().ApiLog(response, requestBody, resource, requestMethod);
+        extentLogger.ApiLog(response, requestBody, resource, requestMethod);
     }
 
     /** Used in Get() ands Delete() to log API info*/
@@ -137,7 +137,7 @@ public class AutoLogger {
         SetUpInfo(requestMethod + ": " + resource);
         SetUpInfo("STATUS CODE: " + response.statusCode());
         SetUpInfo("RESPONSE BODY: " + response.asString());
-        //ExtentFactory.getInstance().ApiLog(response, resource, requestMethod);
+        extentLogger.ApiLog(response, resource, requestMethod);
     }
 
     /** Resets int stepNumber after test method finishes. */
